# Docs/02_contracts_and_schemas/offisim_openapi.yaml
openapi: 3.1.0
info:
  title: Offisim Platform and Local Install APIs
  version: 0.2.0
  summary: Platform marketplace APIs plus local runtime install/import contract snapshots.
  description: >
    This file tracks the current public platform HTTP surface mounted from `apps/platform/src/index.ts`
    together with the local install transaction contract used by `packages/install-core`.
    Public marketplace reads are unauthenticated. Creator publishing, reviews, creator bootstrap,
    install receipt recording, and user library access require authenticated user context.
    Better Auth native session endpoints under `/api/auth/*` and API token management under
    `/v1/auth/tokens` are implementation-owned and intentionally not expanded here.
    Local install/import endpoints remain declaration-driven: no arbitrary install hooks,
    no postinstall scripts, and no remote code execution during import.
servers:
  - url: https://api.offisim.market
    description: Registry and marketplace API
  - url: http://127.0.0.1:43111
    description: Local desktop runtime API
...
security: []
paths:
...
  /v1/install/receipts:
    post:
      tags: [Install Platform]
      summary: Record a successful install receipt on the platform
      operationId: recordInstallReceipt
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallReceiptRequest'
      responses:
        '200':
          description: Install receipt recorded or already present
          content:
            application/json:
              schema:
                type: object
                required: [install_receipt_id, listing_id, package_version_id, status]
                properties:
                  install_receipt_id: { type: string }
                  listing_id: { type: string }
                  package_version_id: { type: string }
                  status:
                    type: string
                    enum: [recorded, already_exists]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/install/download/{versionId}:
    get:
      tags: [Install Platform]
      summary: Get artifact download metadata for a package version
      operationId: getInstallDownload
      security: []
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Artifact metadata
          content:
            application/json:
              schema:
                type: object
                required: [package_version_id, artifact_url]
                properties:
                  package_version_id: { type: string }
                  artifact_url: { type: string, format: uri }
                  artifact_sha256: { type: string, nullable: true }
                  artifact_size_bytes: { type: integer, nullable: true }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token or Better Auth bearer session
  parameters:
...
    VersionId:
      name: versionId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    InstallReceiptRequest:
      type: object
      required: [listing_id, package_version_id, install_source]
      properties:
        listing_id: { type: string }
        package_version_id: { type: string }
        install_source:
          type: string
          enum: [registry, url, file]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true